Pull request #10: Fix model numerical errors when the tempature is too small
Merge in GEN/vllm from chang/fix_temp to feature-based-on-v0.2.4.0

Squashed commit of the following:

commit 391e5040648b72cf577d5c8f6fe4feab4b89e176
Author: Chang Su <chang.s.su@oracle.com>
Date:   Thu Feb 15 13:56:43 2024 -0800

    Fix model numerical errors when the tempature is too small

+++ vllm/model_executor/layers/sampler.py
@@ -63,10 +63,12 @@ class Sampler(nn.Module):
         temperatures = _get_temperatures(sampling_metadata)
         assert len(temperatures) == logits.shape[0]
         if any(t != 1.0 for t in temperatures):
+            # Use float32 for temperature scaling
             t = torch.tensor(temperatures,
-                             dtype=logits.dtype,
+                             dtype=torch.float,
                              device=logits.device)
             # Use in-place division to avoid creating a new tensor.
+            logits = logits.to(torch.float)
             logits.div_(t.unsqueeze(dim=1))
 
         # Apply top-p and top-k truncation.
